Load the package as a SourceFileLoader instead of ExtensionFileLoader
